move existing schemas into frontend/backend folders
