+++ Cargo.toml
@@ -11,6 +11,7 @@ resolver = "2"
 
 [workspace.dependencies]
 futures = "0.3.28"
+glob = "0.3.1"
 serde = { version = "1.0.188", features = ["derive"] }
 anyhow = "1.0.69"
 thiserror = "1.0.48"
