+++ schemas/README.md
@@ -1,0 +1,13 @@
+# Schemas
+
+This directory contains the data models for the frontend and backend.
+
+The schemas are written using [protobuf](https://protobuf.dev/) definitions.
+
+## Backend
+
+The backend schemas reside in the [backend](./protos/backend) folder. The Rust code for each schema is generated by running `cargo build`; it invokes a build script in [build.rs](./build.rs) to compile the proto files into rust.
+
+## Frontend
+
+The frontend schemas reside in the [frontend](./protos/frontend) folder. The JavaScript code for each schema is generated by running `protoc`, which we can run using [proto.sh](../scripts/proto.sh).
