+++ schemas/build.rs
@@ -1,9 +1,14 @@
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let proto_files = vec![
-        "protos/account.proto",
-        "protos/profile.proto",
-        "protos/tweet.proto",
-    ];
+extern crate glob;
+
+use glob::glob;
+use std::error::Error;
+use tonic_build;
+
+fn main() -> Result<(), Box<dyn Error>> {
+    let proto_files: Vec<String> = glob("protos/**/*.proto")?
+        .filter_map(Result::ok)
+        .map(|path| path.display().to_string())
+        .collect();
 
     tonic_build::configure()
         .build_server(true)
