+++ scripts/windows/proto.ps1
@@ -9,4 +9,4 @@ $ErrorActionPreference = "Stop"
     --proto_path='schemas/protos/' `
     --js_out='import_style=commonjs,binary:twote-web/src/proto' `
     --grpc-web_out='import_style=commonjs,mode=grpcwebtext:twote-web/src/proto' `
-    'schemas/protos/*.proto'
+    'schemas/protos/**/*.proto'
